Download source H5AD to local file name.

Download source H5AD to local file name.

Usage

download_source_h5ad(
  dataset_id,
  file,
  overwrite = FALSE,
  census_version = "stable",
  census = NULL
)

Arguments

dataset_id

The dataset_id of interest.

file

Local file name to store H5AD file.

overwrite

TRUE to allow overwriting an existing file.

census_version

The desired Census version.

census

An open Census handle for census_version. If not provided, then it will be opened and closed automatically; but it's more efficient to reuse a handle if calling download_source_h5ad() multiple times.

Examples

download_source_h5ad("0895c838-e550-48a3-a777-dbcd35d30272", "/tmp/data.h5ad", overwrite = TRUE)